+++ node/clients/search.ts
@@ -308,13 +308,14 @@ export class Search extends AppClient {
       metric: 'search-category',
     })
 
-  public crossSelling = (id: string, type: SearchCrossSellingTypes) =>
-    this.get<SearchProduct[]>(
-      `/pub/products/crossselling/${type}/${id}?groupByProduct=true`,
+  public crossSelling = (id: string, type: SearchCrossSellingTypes, groupByProduct = true) =>{
+    return this.get<SearchProduct[]>(
+      `/pub/products/crossselling/${type}/${id}?groupByProduct=${groupByProduct}`,
       {
         metric: 'search-crossSelling',
       }
     )
+  }
 
   public filtersInCategoryFromId = (id: string | number) =>
     this.get<FilterListTreeCategoryById[]>(